Working game engine, implementing power ups

+++ LudumDare29/TextureManager.h
@@ -24,8 +24,8 @@ public:
     static TextureManager* Instance();
     
     bool load(string fileName, string textureID, SDL_Renderer* pRenderer);
-    void draw(string textureID, int x, int y,
-              int width, int height, SDL_Renderer* pRenderer);
+    void draw(string textureID, float x, float y,
+              float width, float height, SDL_Renderer* pRenderer);
     
     void clearTextureMap();
     void clean();
